# copr: improve like algorithm performance