--- a/components/tidb_query/src/expr_util/like.rs
+++ b/components/tidb_query/src/expr_util/like.rs
@@ -14,75 +14,49 @@ use std::slice::Iter;
 
 use crate::expr::Result;
 
-const MAX_RECURSE_LEVEL: usize = 1024;
-
-// Do match until '%' is found.
-#[inline]
-fn partial_like(tcs: &mut Iter<'_, u8>, pcs: &mut Iter<'_, u8>, escape: u32) -> Option<bool> {
-    loop {
-        match pcs.next().cloned() {
-            None => return Some(tcs.next().is_none()),
-            Some(b'%') => return None,
-            Some(c) => {
-                let (npc, escape) = if u32::from(c) == escape {
-                    pcs.next().map_or((c, false), |&c| (c, true))
-                } else {
-                    (c, false)
-                };
-                let nsc = match tcs.next() {
-                    None => return Some(false),
-                    Some(&c) => c,
-                };
-                if nsc != npc && (npc != b'_' || escape) {
-                    return Some(false);
-                }
-            }
-        }
-    }
-}
-
-pub fn like(target: &[u8], pattern: &[u8], escape: u32, recurse_level: usize) -> Result<bool> {
-    let mut tcs = target.iter();
-    let mut pcs = pattern.iter();
-    loop {
-        if let Some(res) = partial_like(&mut tcs, &mut pcs, escape) {
-            return Ok(res);
-        }
-        let next_char = loop {
-            match pcs.next().cloned() {
-                Some(b'%') => {}
-                Some(b'_') => {
-                    if tcs.next().is_none() {
-                        return Ok(false);
+pub fn like(target: &[u8], pattern: &[u8], escape: u32) -> Result<bool> {
+    let mut px = 0;
+    let mut tx = 0;
+    let mut next_px = 0;
+    let mut next_tx = 0;
+    while px < pattern.len() || tx < target.len() {
+        if px < pattern.len() {
+            let c = pattern[px];
+            match c {
+                b'_' => {
+                    if tx < target.len() {
+                        px += 1;
+                        tx += 1;
+                        continue;
                     }
                 }
-                // So the pattern should be some thing like 'xxx%'
-                None => return Ok(true),
-                Some(c) => {
-                    break if u32::from(c) == escape {
-                        pcs.next().map_or(escape, |&c| u32::from(c))
-                    } else {
-                        u32::from(c)
-                    };
+                b'%' => {
+                    next_px = px;
+                    next_tx = tx + 1;
+                    px += 1;
+                    continue;
+                }
+                mut pc => {
+                    if pc as u32 == escape && px + 1 < pattern.len() {
+                        px += 1;
+                        pc = pattern[px];
+                    }
+                    if tx < target.len() && target[tx] == pc {
+                        tx += 1;
+                        px += 1;
+                        continue;
+                    }
                 }
             }
-        };
-        if recurse_level >= MAX_RECURSE_LEVEL {
-            // TODO: maybe we should test if stack is actually about to overflow.
-            return Err(box_err!(
-                "recurse level should not be larger than {}",
-                MAX_RECURSE_LEVEL
-            ));
         }
-        // Pattern must be something like "%xxx".
-        loop {
-            let s = match tcs.next() {
-                None => return Ok(false),
-                Some(&s) => u32::from(s),
-            };
-            if s == next_char && like(tcs.as_slice(), pcs.as_slice(), escape, recurse_level + 1)? {
-                return Ok(true);
-            }
+        // Mismatch and backtrace.
+        if 0 < next_tx && next_tx <= target.len() {
+            px = next_px;
+            tx = next_tx;
+            continue;
         }
+        return Ok(false);
     }
+
+    Ok(px == pattern.len())
 }