--- a/components/tidb_query/src/rpn_expr/impl_like.rs
+++ b/components/tidb_query/src/rpn_expr/impl_like.rs
@@ -14,12 +14,11 @@ pub fn like(
     escape: &Option<i64>,
 ) -> Result<Option<i64>> {
     match (target, pattern, escape) {
-        (Some(target), Some(pattern), Some(escape)) => {
-            Ok(Some(
-                expr_util::like::like(target.as_slice(), pattern.as_slice(), *escape as u32)?
-                    as i64,
-            ))
-        }
+        (Some(target), Some(pattern), Some(escape)) => Ok(Some(expr_util::like::like(
+            target.as_slice(),
+            pattern.as_slice(),
+            *escape as u32,
+        )? as i64)),
         _ => Ok(None),
     }
 }
@@ -63,7 +62,12 @@ mod tests {
             (r#"3hello"#, r#"3%hello"#, '3', Some(0)),
             (r#"3hello"#, r#"__hello"#, '_', Some(0)),
             (r#"3hello"#, r#"%_hello"#, '%', Some(1)),
-            (r#"aaaaaaaaaaaaaaaaaaaaaaaaaaa"#, r#"a%a%a%a%a%a%a%a%b"#, '\\', Some(0)),
+            (
+                r#"aaaaaaaaaaaaaaaaaaaaaaaaaaa"#,
+                r#"a%a%a%a%a%a%a%a%b"#,
+                '\\',
+                Some(0),
+            ),
         ];
         for (target, pattern, escape, expected) in cases {
             let output = RpnFnScalarEvaluator::new()