PCP #5709
 Signed-off-by: TennyZhuang zty0826@gmail.com

<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.
-->

###  What have you changed?

#5709

Please explain in detail what the changes are in this PR and why they are needed:

- Optimize like() implementation by a better like algorithm to avoid common glob attacks, described in https://research.swtch.com/glob.
- The main idea is backtrace always return to last `%`, so that every character will be traversed at most twice.
- Optimize the worst time complexity of `like(T, P)` from `O(2^len(P))` to `len(P) + len(T)`.

###  What is the type of the changes?

Pick one of the following and delete the others:

- Improvement (a change which is an improvement to an existing feature)

###  How is the PR tested?
Please select the tests that you ran to verify your changes:
- Unit test
- Manual test (add detailed scripts or steps below)

Some benchmark code in https://github.com/TennyZhuang/like-bench/

###  Does this PR affect documentation (docs) or should it be mentioned in the release notes?

###  Does this PR affect `tidb-ansible`?

No

###  Refer to a related PR or issue link (optional)

#5709

###  Benchmark result if necessary (optional)

A little overhead on naive pattern, but some optimization on complicated pattern and huge improvement on attack.

E.g.:

For input `(r#"Hello, World"#, r#"Hello, %"#, '\\', Some(1))`

```
like 2                  time:   [12.410 ns 12.462 ns 12.525 ns]
Found 19 outliers among 100 measurements (19.00%)
  3 (3.00%) high mild
  16 (16.00%) high severe

like_optimize 2         time:   [15.834 ns 15.979 ns 16.184 ns]
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe
```

A little overhead.

And for input `(r#"3hello"#, r#"%_hello"#, '%', Some(1))`

```
like 29                 time:   [14.004 ns 14.573 ns 15.403 ns]
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe

like_optimize 29        time:   [9.7671 ns 9.8921 ns 10.031 ns]
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
```

about 50% improvement.

The most important is glob attack, for a special constructed input (target="a"*100, pattern="a%"*n+b), original like implementation will produce a terrible result.

![31573469938_ pic](https://user-images.githubusercontent.com/9161438/68641484-f3d61f80-0545-11ea-9551-48a487040c9a.jpg)


And after optimization, the cost is almost linear even on large input size.

![51573470753_ pic](https://user-images.githubusercontent.com/9161438/68641490-fb95c400-0545-11ea-8141-b3e514f03e41.jpg)


Please note that the unit of y axis in first graph is ms and in second graph is ns.

All benchmark can be reproduced by run `cargo bench` in https://github.com/TennyZhuang/like-bench/

###  Any examples? (optional)