@sticnarf 2019-11-12
By the way, both the original implementation and the implementation in this PR don't support wide characters like 汉字. 
@breeswish Do you have any ideas about dealing with the problem? The like function should be aware the `Bytes` is a string or a byte array.