@TennyZhuang 2019-11-12
@sticnarf I think it can be fixed by iterate per char instead of per byte, but it should not be done in one PR, it will affect the benchmark.