@breeswish 2019-11-12
@sticnarf Looks like MySQL also handles byte by byte:

https://stackoverflow.com/questions/6482531/how-to-make-mysql-aware-of-multi-byte-characters-in-like-and-regexp