@breeswish 2019-11-12
Update: I discovered that MySQL's LIKE both supports binary mode and character mode. Currently we only implemented the binary mode like. I think it's fine to only provide a binary mode like in this PR. However it would be better to support both.