@TennyZhuang 2019-11-12
I'd like to fix it in another PR, and I think the char mode should be implemented in TiDB firstly (I'd like to port the new implementation to TiDB).