@sticnarf 2019-11-13
The algorithm looks good to me. But it'd be better if you can add some comments to the variables used in the algorithm to make it more understandable.