@TennyZhuang 2019-11-13
@sticnarf please take a look at the before implementation and the testcase, it seems that the precedence of escape is lower than wildcard.