@sticnarf 2019-11-13
> @sticnarf please take a look at the before implementation and the testcase, it seems that the precedence of escape is lower than wildcard.

Oh really sorry. You're right, keep it as it is now.

But it's quite an awkward situation... MySQL 5.7 thinks the precendence of escape is lower, however MySQL 8.0 and MariaDB 10.4 think it's higher...