@TennyZhuang 2019-11-13
<img width="934" alt="image" src="https://user-images.githubusercontent.com/9161438/68763946-aa282a80-0654-11ea-99b5-1341d3ef63eb.png">

Why DCO check blocks on waiting